t01/src/requirements.txt
pytest
python-dotenv
python-telegram-bot==20.0
pytest-asyncio
